requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.6.0
deep-translator==1.11.4
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
kafka-python-ng==2.2.3
MarkupSafe==2.1.5
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
pymongo==4.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
